+++ src/namespaces/dec112.ts
@@ -1,4 +1,3 @@
-import { IncomingMessage } from 'jssip/lib/SIPMessage';
 import { CALL_INFO, CONTENT_TRANSFER_ENCODING, CONTENT_TYPE } from '../constants/headers';
 import { X_DEC112_TEST, X_DEC112_TEST_VALUE_TRUE } from '../constants/headers/dec112';
 import { fromEmergencyMessageType, toEmergencyMessageType } from '../constants/message-types/dec112';
@@ -6,6 +5,7 @@ import { ConversationEndpointType } from '../models/conversation';
 import type { PidfLo } from 'pidf-lo'
 import { EmergencyMapper, getRegEx, regexHeaders } from './emergency';
 import { MessageParts, MessagePartsParams, NamespacedConversation, NamespaceSpecifics } from './interfaces';
+import { NewMessageEvent } from '../models/sip-agent';
 import { Base64 } from '../utils/base64';
 
 const dec112Domain = 'service.dec112.at';
@@ -139,7 +139,7 @@ export class DEC112Mapper implements NamespacedConversation {
   getCallIdFromHeaders = (headers: string[]): string | undefined => regexHeaders(headers, getRegEx(getCallIdHeaderValue));
   getMessageIdFromHeaders = (headers: string[]): string | undefined => regexHeaders(headers, getRegEx(getMessageIdHeaderValue));
   getDIDFromHeaders = (headers: string[]): string | undefined => EmergencyMapper.getDIDFromHeaders(headers);
-  getIsTestFromHeaders = (message: IncomingMessage): boolean => message.getHeader(X_DEC112_TEST) === X_DEC112_TEST_VALUE_TRUE;
+  getIsTestFromHeaders = (message: NewMessageEvent): boolean => message.getHeader(X_DEC112_TEST) === X_DEC112_TEST_VALUE_TRUE;
   getMessageTypeFromHeaders = (headers: string[], messageText?: string): number | undefined => {
     const msgType = regexHeaders(headers, getRegEx(getMessageTypeHeaderValue));