+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  
 ## [Unreleased]
+
+---
+## Unreleased - [2.0.0]
 ### Added
 - Agent mode that can be used to indicate an app is running in the background
 - Support for DIDs.
+- Support for binaries
 
 ---