+++ CHANGELOG.md
@@ -8,10 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 ## Unreleased - [2.0.0]
+### Changed
+- **BREAKING**: `JsSIP` is no longer a direct dependency, but a peer dependency of `ng112-js`. \
+This is due to the added compatibility with `sip.js`. \
+Developers now can choose which SIP library they want to use by just installing the preferred one. \
+If both SIP libraries are installed, developers can define property `preferredSipAgent` when creating the ng112-js `Agent`. \
+If nothing is specified, ng112-js will fall back to `JsSIP`.
+- **BREAKING**: New object for message send errors: `MessageError` instead of former object `MessageFailedEvent` \
+Object has different properties!
 ### Added
-- Agent mode that can be used to indicate an app is running in the background
+- Support for `sip.js` in addition to `JsSIP` as SIP stack.
+- Agent mode that can be used to indicate an app is running in the background.
 - Support for DIDs.
 - Support for binaries
+- Support 
 
 ---
 
@@ -26,7 +36,9 @@ This allows for the use of `extraParts`, `uris` and other properties for `start`
 - **BREAKING**: The agent's `conversation` property contains only calls that are not in state `STOPPED`. \
 Once calls are stopped they are removed from this list. \
 This should limit memory consumption if ng112-js is used in long running applications with heavy load.
-
+- The agent's `conversation` property only contains open calls. Stopped calls are removed from this list. \
+`debug` will now also accept a callback function that is called for all log messages. \
+This allows for the connection of any other logging facility.
 ### Added
 - Property `uniqueId` in messages that is unique across all messages and conversations within ng112-js
 - Auto-close of open calls on agent disposal